CLDR-5973 Move some code from XMLSource to new class CodeFallback (#4420)
